requirements.txt
h5py==2.10.0
numpy==1.19.1
Pillow==7.2.0
scipy==1.5.2
six==1.15.0
